+++ source/dotnet/wholesale-api/FunctionApp.IntegrationTests/Orchestration.IntegrationTests.csproj
@@ -1,8 +1,8 @@
 ﻿<Project Sdk="Microsoft.NET.Sdk">
 
   <PropertyGroup>
-    <AssemblyName>Energinet.DataHub.Wholesale.FunctionApp.IntegrationTests</AssemblyName>
-    <RootNamespace>Energinet.DataHub.Wholesale.FunctionApp.IntegrationTests</RootNamespace>
+    <AssemblyName>Energinet.DataHub.Wholesale.Orchestration.IntegrationTests</AssemblyName>
+    <RootNamespace>Energinet.DataHub.Wholesale.Orchestration.IntegrationTests</RootNamespace>
     <IsPackable>false</IsPackable>
     <IsTestProject>true</IsTestProject>
   </PropertyGroup>
@@ -22,7 +22,7 @@
   </ItemGroup>
 
   <ItemGroup>
-    <ProjectReference Include="..\FunctionApp\FunctionApp.csproj" />
+    <ProjectReference Include="..\FunctionApp\Orchestration.csproj" />
   </ItemGroup>
 
   <ItemGroup>
