+++ source/dotnet/wholesale-api/FunctionApp/Monitor/HealthCheckEndpoint.cs
@@ -16,7 +16,7 @@ using Energinet.DataHub.Core.App.FunctionApp.Diagnostics.HealthChecks;
 using Microsoft.Azure.Functions.Worker;
 using Microsoft.Azure.Functions.Worker.Http;
 
-namespace Energinet.DataHub.Wholesale.FunctionApp.Monitor
+namespace Energinet.DataHub.Wholesale.Orchestration.Monitor
 {
     public class HealthCheckEndpoint
     {
